+++ packages/coding-agent/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Added `auth.accountPriority` setting and `/account priority` command (also `/session priority`) to set and manage priority order for providers with multiple accounts.
+- Added `auth.accountPriority` setting and `/account priority` command (also `/session priority`) to set and manage priority order for providers with multiple accounts ([#11645](https://github.com/can1357/oh-my-pi/pull/11645) by [@keethesh](https://github.com/keethesh)).
 
 ## [18.1.17] - 2026-09-10